Rename data-extension-editor files.
