+++ extensions/ql-vscode/src/model-editor/model-editor-view.ts
